refactor: avoid creating new procesor instances each time one is needed
