+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/dependent/kubernetes/processors/ConfigMapResourceUpdatePreProcessor.java
@@ -1,0 +1,16 @@
+package io.javaoperatorsdk.operator.processing.dependent.kubernetes.processors;
+
+import io.fabric8.kubernetes.api.model.ConfigMap;
+import io.javaoperatorsdk.operator.processing.dependent.kubernetes.ResourceUpdatePreProcessor;
+
+public class ConfigMapResourceUpdatePreProcessor
+    extends GenericResourceUpdatePreProcessor<ConfigMap> {
+  final static ResourceUpdatePreProcessor<ConfigMap> INSTANCE =
+      new ConfigMapResourceUpdatePreProcessor();
+
+  @Override
+  protected void updateClonedActual(ConfigMap actual, ConfigMap desired) {
+    actual.setData(desired.getData());
+    actual.setBinaryData((desired.getBinaryData()));
+  }
+}
